fix(rust): move split_to_morsels to shared so ipc builds without parquet
ipc imported split_to_morsels from the parquet module, so a feature set with ipc and without parquet failed to compile. Relocate the helper next to the other shared io-source utilities.

Rebased onto current main after #27898 was confirmed still open on 0.55.2.
