+++ crates/polars-stream/src/nodes/io_sources/parquet/init.rs
@@ -2,7 +2,6 @@ use std::sync::Arc;
 
 use arrow::datatypes::ArrowDataType;
 use polars_async::executor;
-use polars_core::frame::DataFrame;
 use polars_core::runtime::ASYNC;
 use polars_error::{PolarsResult, polars_ensure};
 use polars_io::prelude::_internal::PrefilterMaskSetting;
@@ -16,6 +15,7 @@ use crate::morsel::{Morsel, SourceToken, get_ideal_morsel_size};
 use crate::nodes::io_sources::multi_scan::reader_interface::output::FileReaderOutputSend;
 use crate::nodes::io_sources::parquet::projection::ArrowFieldProjection;
 use crate::nodes::io_sources::parquet::statistics::calculate_row_group_pred_pushdown_skip_mask;
+use crate::nodes::io_sources::shared::morsel_split::split_to_morsels;
 use crate::nodes::{MorselSeq, TaskPriority};
 use crate::utils::tokio_handle_ext::{self, AbortOnDropHandle};
 
@@ -405,31 +405,6 @@ fn filtered_range(exclude: &[usize], len: usize) -> impl Iterator<Item = usize> 
             false
         }
     })
-}
-
-pub(crate) fn split_to_morsels(
-    df: &DataFrame,
-    ideal_morsel_size: usize,
-    last_morsel: bool,
-    last_morsel_pipelines: usize,
-) -> impl Iterator<Item = DataFrame> + '_ {
-    let mut n_morsels = if df.height() > 3 * ideal_morsel_size / 2 {
-        // num_rows > (1.5 * ideal_morsel_size)
-        (df.height() / ideal_morsel_size).max(2)
-    } else {
-        1
-    };
-
-    if last_morsel {
-        n_morsels = n_morsels.max(last_morsel_pipelines);
-    }
-
-    let rows_per_morsel = df.height().div_ceil(n_morsels).max(1);
-
-    (0..i64::try_from(df.height()).unwrap())
-        .step_by(rows_per_morsel)
-        .map(move |offset| df.slice(offset, rows_per_morsel))
-        .filter(|df| df.height() > 0)
 }
 
 mod tests {
