+++ crates/polars-stream/src/nodes/io_sources/shared/mod.rs
@@ -1,4 +1,6 @@
 #[cfg(any(feature = "csv", feature = "json", feature = "scan_lines"))]
 pub mod chunk_data_fetch;
 #[cfg(any(feature = "parquet", feature = "ipc"))]
+pub mod morsel_split;
+#[cfg(any(feature = "parquet", feature = "ipc"))]
 pub mod pipeline_budget;
